remove EOF token
The eof token cannot be generated in the lexer without generating an invalid span.
This is especially true for a valid file, since you cannot use the span of
the last character in the input (since an empty file does not have a character and
therefore does not have a valid span.)

Removing the EOF token does not only simplify the logic, but also removes the
invalid internal state of "MissingEOF token" that should never occur.