fix #74: take dyn Display instead of dyn AsFail in Diagnostics Api
Fail has Display as Trait bound, and we only use the Display API, not
any of the Fail APIs. Since Diagnostics was extended to writing info
messages in addition to error messages, changing the trait bound does
not only make sense from a technial, but also a semantic view point.