mjrt: fix handling of zero-sized-types inmjrt_new
calloc might return NULL if called with size = 0.
We cannot let this happen since mjrt_new is used for
MiniJava `new`, which must work for 0-sized classes.
This means specifically that a `==` comparison of two
objects of the same class must return false, which does
not work if we return NULL here.