lir: CopyPropagation: flow params (CopyPropagationSrc) (fixup 9198e48ca)
Given the current implementation of lva / regalloc, the following
example uses less registers for both SysVAMD64 ("x86") and cdecl
("stack") calling convention:

 public int if_diamond(boolean b, int x, int s, int r) {
     int y;
     if (true == b) {
         y = s;
     } else {
         y = r;
     }
     return y + x;
 }

For a more homogenous design, we should have a LoadParam{idx, dst_slot}
instruction for the LIR, and each start block should load all Params
into value slots.  There is no information loss toward the register
allocation, because LoadParam{idx, dst_slot} embodies the information
that dst_slot stems from a parameter (and thus may be pre-allocated with
the cconv's register, or may not need to be spilled because it can be
loaded from the stack).

A note about the following kind of change:

 -            if !src.firm().mode().is_mem() && !dst.firm.mode().is_mem() {
 -                instrs.push(Instruction::Mov {
 -                    src: *src,
 -                    dst: dst.multislot(),
 -                });
 -            }
 +            instrs.push(Instruction::Mov {
 +                src: *src,
 +                dst: dst.multislot(),
 +            })

It is safe to drop the check for is_mem here since LIR guarantees us
that no mem values will be flowed (i.e. copy propagated)