lir: IMPROVE THIS: fix missing copy propagation for phi-to-phi
Test that uncovered this: insertion_sort.java

The j = j-1 from the inner loop is stored on the stack.
And the entry of the inner loop header loads j from there.
But the initial value for j, that is, int j = i in the outer loop body,
is itself a Phi node of i={0,i+1}.

It seems like add_incoming_value_flow's check for whether a target slot
for the outer Phi node already exists hits too often, i.e. it assumes
the outer Phi value is already being flowed to the block where the inner
Phi value is defined (PIGGY is emitted in that case).

The "fix" is to check whether we are in this Phi-to-Phi case.
But I'd rather see a systematic investigation why we detect this
'already being flowed' case wrong for Phi nodes.

Spots to check:

MultiSlotBuilder commits on every add_possible_value, whereas I'd expect
it to only commit on get_multislot.

@joshuabach your turn :P