lir: handle sel & member nodes that are used in another block correctly
Since we use sel.ptr() and member.ptr(), we must flow _those_ nodes
instead of nothing at all (that was the previous situation) or the sel
nodes themselves (sel nodes only produce AddressComputation, hence no
ValueSlot).

Spotted with the safety feature enabled + the folloing code:

class Main {
      public static void main(String[] args) {
          int[] foo = new int[5];

          System.out.println(foo[-5236790]);
      }
}

In the FIRM graph, you can see that the Sel for foo[...] is in the
start block, but the Sel is only used by a subsequent Load if the
null check passes.

@joshuabach please convert this to a test case