copy prop sort: always use 8 Byte registers
We can't know the actual size of the registers when we transform
`RegToRegTransfer` back into Instructions, since we insert the
`DstOperand`s, ignoring their size into a `HashMap`. This results in
following behavior:
We get following from the lir/lva:
```
%edx -> %edi
%rdi -> %rdx
```
We insert the registers in the order `%edx`, `%edi`, `%rdi` and `%rdx`
where the last 2 won't get inserted, because their smaller counterpart
is already in the `HashMap`.

Since this is a 2-cycle, it doesn't/shouldn't matter how we resolve
this. However, if the algorithm decides to resolve this, by removing
the edge `%edx -> %edi` (push on stack) the following code will be
emitted:
```
/* The 8 byte variant, since we always need to push 8 bytes on the
   stack */
pushq %rdx
/* Move of 4 byte registers, since we added those as 4 byte registers.
   The problem lays here, since those registers will be cut off */
movl %edi, %edx
/* Pop %rdx into %rsi */
popq %rsi
```